+++ src/main/java/com/drzewo97/ballotbox/core/model/election/ElectionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ElectionRepository extends CrudRepository<Election, Long> {
+public interface ElectionRepository extends CrudRepository<Election, Integer> {
 	Boolean existsByName(String name);
 }
