+++ src/main/java/com/drzewo97/ballotbox/core/service/committee/CommitteeService.java
@@ -1,5 +1,5 @@
 package com.drzewo97.ballotbox.core.service.committee;
 
 public interface CommitteeService {
-	Boolean isCommitteeAdmin(String username, Long wardId);
+	Boolean isCommitteeAdmin(String username, Integer wardId);
 }
