+++ src/main/java/com/drzewo97/ballotbox/panel/controller/poll/PollResultController.java
@@ -41,7 +41,7 @@ public class PollResultController {
 	private CandidatesVotesCountWardProtocolRepository wardProtocolRepository;
 	
 	@GetMapping
-	public String pollResult(Model model, @PathVariable("id") Long id){
+	public String pollResult(Model model, @PathVariable("id") Integer id){
 		Optional<Poll> poll = pollRepository.findById(id);
 		if(poll.isEmpty()){
 			return "redirect:/poll/all?empty";
